requirements.txt
numpy<2.0.0
chromadb
protobuf==3.20.3
pandas==1.5.3
scikit-learn==1.2.2
langchain
langchain-community
langchain-groq
unstructured
selenium
python-dotenv==1.0.0
fastapi
uvicorn[standard]
jinja2
pytest
httpx
